Agregando carpeta impl dentro de repository (logica)

+++ sistema-gestion-biblioteca/src/main/java/modelo/Usuario.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 
 public class Usuario {
-    private int dni;
+    private Long dni;
     private String nombre;
     private String email;
     private EstadoUsuario estadoUsuario;
